+++ src/components/TaskSection.tsx
@@ -1,14 +1,7 @@
 "use client";
 
-import {
-    Button,
-    Group,
-    Paper,
-    ScrollAreaAutosize,
-    Stack,
-    Title,
-} from "@mantine/core";
-import { TaskCard } from "./TaskCard";
+import { Group, Paper, ScrollAreaAutosize, Stack, Title } from "@mantine/core";
+import { TaskCard } from "./TaskCard/TaskCard";
 import { useDroppable } from "@dnd-kit/core";
 import { Section as SectionType, Task } from "types/tasks";
 import {
@@ -16,15 +9,23 @@ import {
     verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
 import { memo } from "react";
+import { RiAddLine } from "@remixicon/react";
+import { NewTaskButton } from "./NewTaskButton";
 
 type TaskSectionProps = {
     section: SectionType;
     tasks: Task[];
     handleAddTask: (sectionId: SectionType["id"]) => void;
+    handleContentChange: (task: Task, content: string) => void;
 };
 
 export const TaskSection = memo(
-    ({ section, tasks, handleAddTask }: TaskSectionProps) => {
+    ({
+        section,
+        tasks,
+        handleAddTask,
+        handleContentChange,
+    }: TaskSectionProps) => {
         const { setNodeRef } = useDroppable({
             id: section.id,
         });
@@ -44,9 +45,11 @@ export const TaskSection = memo(
                             {section.title}
                         </Title>
                         {section.id === "IN_PROGRESS" && (
-                            <Button onClick={() => handleAddTask(section.id)}>
-                                Add
-                            </Button>
+                            <NewTaskButton
+                                onClick={() => handleAddTask(section.id)}
+                            >
+                                <RiAddLine />
+                            </NewTaskButton>
                         )}
                     </Group>
 
@@ -64,6 +67,9 @@ export const TaskSection = memo(
                                     <TaskCard
                                         key={task.id}
                                         task={task}
+                                        handleContentChange={
+                                            handleContentChange
+                                        }
                                     ></TaskCard>
                                 ))}
                             </Stack>